+++ int-ui/int-ui-standalone/src/main/kotlin/org/jetbrains/jewel/intui/standalone/styling/IntUiTooltipStyling.kt
@@ -1,20 +1,13 @@
 package org.jetbrains.jewel.intui.standalone.styling
 
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.DpOffset
-import androidx.compose.ui.unit.dp
+import org.jetbrains.jewel.intui.core.styling.defaults
 import org.jetbrains.jewel.intui.core.theme.IntUiDarkTheme
 import org.jetbrains.jewel.intui.core.theme.IntUiLightTheme
 import org.jetbrains.jewel.styling.TooltipColors
 import org.jetbrains.jewel.styling.TooltipMetrics
 import org.jetbrains.jewel.styling.TooltipStyle
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.milliseconds
 
 @Composable
 fun TooltipStyle.Companion.light(
@@ -49,13 +42,3 @@ fun TooltipColors.Companion.dark(
     borderColor: Color = IntUiDarkTheme.colors.grey(3),
     shadow: Color = Color(0x66000000), // Not a palette color
 ) = TooltipColors(backgroundColor, contentColor, borderColor, shadow)
-
-fun TooltipMetrics.Companion.defaults(
-    contentPadding: PaddingValues = PaddingValues(vertical = 9.dp, horizontal = 12.dp),
-    showDelay: Duration = 0.milliseconds,
-    cornerSize: CornerSize = CornerSize(5.dp),
-    borderWidth: Dp = 1.dp,
-    shadowSize: Dp = 12.dp,
-    tooltipOffset: DpOffset = DpOffset(0.dp, 20.dp),
-    tooltipAlignment: Alignment.Horizontal = Alignment.Start,
-) = TooltipMetrics(contentPadding, showDelay, cornerSize, borderWidth, shadowSize, tooltipOffset, tooltipAlignment)
